Added bottom title padding to the dialogs
